Create CardGame.java
A program on which the user inserts three cards. The program then verifies if they are indeed possible cards or just a random String. If they are all true, it outputs 5 different things: 

1. Are they all of the same kind?
2. Are they all from different kinds?
3. Do they form a three of a kind?
4. Do they form a flush?
5. Do they form a straight flush?

It was kinda challenging and really fun to do, helping a lot with String and char manipulation.